[SMAR-3101] - unificação chamadas de pagamento
