--- app/src/main/java/br/com/uol/pagseguro/smartcoffee/demo/Transaction.java
@@ -1,25 +1,0 @@
-package br.com.uol.pagseguro.smartcoffee.demo;
-
-public class Transaction {
-
-    String transactionId;
-
-    String TransactionCode;
-
-    public String getTransactionId() {
-        return transactionId;
-    }
-
-    public void setTransactionId(String transactionId) {
-        this.transactionId = transactionId;
-    }
-
-    public String getTransactionCode() {
-        return TransactionCode;
-    }
-
-    public void setTransactionCode(String transactionCode) {
-        TransactionCode = transactionCode;
-    }
-}
-
