+++ app/src/main/java/br/com/uol/pagseguro/smartcoffee/demoInterno/DemoInternoContract.java
@@ -1,4 +1,4 @@
-package br.com.uol.pagseguro.smartcoffee.demo;
+package br.com.uol.pagseguro.smartcoffee.demoInterno;
 
 import com.hannesdorfmann.mosby.mvp.MvpView;
