+++ app/src/main/java/br/com/uol/pagseguro/smartcoffee/demoInterno/DismissListener.java
@@ -1,4 +1,4 @@
-package br.com.uol.pagseguro.smartcoffee.demo;
+package br.com.uol.pagseguro.smartcoffee.demoInterno;
 
 public interface DismissListener {
