+++ app/src/main/java/br/com/uol/pagseguro/smartcoffee/injection/DemoInternoComponent.java
@@ -1,7 +1,7 @@
 package br.com.uol.pagseguro.smartcoffee.injection;
 
-import br.com.uol.pagseguro.smartcoffee.demo.DemoInternoActivity;
-import br.com.uol.pagseguro.smartcoffee.demo.DemoInternoPresenter;
+import br.com.uol.pagseguro.smartcoffee.demoInterno.DemoInternoActivity;
+import br.com.uol.pagseguro.smartcoffee.demoInterno.DemoInternoPresenter;
 import dagger.Component;
 
 @Component(modules = {UseCaseModule.class, WrapperModule.class})
