--- app/src/main/java/br/com/uol/pagseguro/smartcoffee/transactions/TransactionsUseCase.java
@@ -1,223 +1,0 @@
-package br.com.uol.pagseguro.smartcoffee.transactions;
-
-
-import java.util.Locale;
-import java.util.Random;
-
-import br.com.uol.pagseguro.plugpagservice.wrapper.PlugPag;
-import br.com.uol.pagseguro.plugpagservice.wrapper.PlugPagAbortResult;
-import br.com.uol.pagseguro.plugpagservice.wrapper.PlugPagCustomPrinterLayout;
-import br.com.uol.pagseguro.plugpagservice.wrapper.PlugPagPaymentData;
-import br.com.uol.pagseguro.plugpagservice.wrapper.PlugPagPrintResult;
-import br.com.uol.pagseguro.plugpagservice.wrapper.PlugPagPrinterListener;
-import br.com.uol.pagseguro.plugpagservice.wrapper.PlugPagTransactionResult;
-import br.com.uol.pagseguro.plugpagservice.wrapper.PlugPagVoidData;
-import br.com.uol.pagseguro.plugpagservice.wrapper.exception.PlugPagException;
-import br.com.uol.pagseguro.smartcoffee.ActionResult;
-import static br.com.uol.pagseguro.smartcoffee.utils.InstallmentConstants.*;
-import static br.com.uol.pagseguro.smartcoffee.utils.SmartCoffeeConstants.*;
-import io.reactivex.Observable;
-import io.reactivex.ObservableEmitter;
-
-public class TransactionsUseCase {
-
-    private final PlugPag mPlugPag;
-    private PlugPagPaymentData mPlugPagPaymentData = null;
-
-    public TransactionsUseCase(PlugPag plugPag) {
-        mPlugPag = plugPag;
-    }
-
-    public Observable<ActionResult> doCreditPayment() {
-        return doPayment(new PlugPagPaymentData(
-                TYPE_CREDITO,
-                getAmount(),
-                INSTALLMENT_TYPE_A_VISTA,
-                1,
-                USER_REFERENCE,
-                true
-        ));
-    }
-
-    public Observable<ActionResult> doCreditPaymentWithSellerInstallments() {
-        return doPayment(new PlugPagPaymentData(
-                TYPE_CREDITO,
-                getAmount(),
-                INSTALLMENT_TYPE_PARC_VENDEDOR,
-                getInstallments(),
-                USER_REFERENCE,
-                true));
-    }
-
-    public Observable<ActionResult> doCreditPaymentWithBuyerInstallments() {
-        return doPayment(new PlugPagPaymentData(
-                TYPE_CREDITO,
-                getAmount(),
-                INSTALLMENT_TYPE_PARC_COMPRADOR,
-                getInstallments(),
-                USER_REFERENCE,
-                true));
-    }
-
-    public Observable<ActionResult> doDebitPayment() {
-        return doPayment(new PlugPagPaymentData(
-                TYPE_DEBITO,
-                getAmount(),
-                INSTALLMENT_TYPE_A_VISTA,
-                1,
-                USER_REFERENCE,
-                true));
-    }
-
-    public Observable<ActionResult> doVoucherPayment() {
-        return doPayment(new PlugPagPaymentData(
-                TYPE_VOUCHER,
-                getAmount(),
-                INSTALLMENT_TYPE_A_VISTA,
-                1,
-                USER_REFERENCE,
-                true));
-    }
-
-    public Observable<ActionResult> doRefundPayment(ActionResult actionResult) {
-        if (actionResult.getTransactionCode() == null) {
-            return Observable.error(new Exception("Nenhuma transação encontrada"));
-        }
-        return doRefund(new PlugPagVoidData(actionResult.getTransactionCode(), actionResult.getTransactionId(), true));
-    }
-
-    private Observable<ActionResult> doRefund(final PlugPagVoidData plugPagVoidData) {
-        return Observable.create(emitter -> {
-            ActionResult result = new ActionResult();
-            setListener(emitter, result);
-            setPrintListener(emitter, result);
-            mPlugPag.setPlugPagCustomPrinterLayout(getCustomPrinterDialog());
-            PlugPagTransactionResult plugPagTransactionResult = mPlugPag.voidPayment(plugPagVoidData);
-            sendResponse(emitter, plugPagTransactionResult, result);
-        });
-    }
-
-    private Observable<ActionResult> doPayment(final PlugPagPaymentData paymentData) {
-        mPlugPagPaymentData = paymentData;
-        return Observable.create(emitter -> {
-            mPlugPag.setPlugPagCustomPrinterLayout(getCustomPrinterDialog());
-            ActionResult result = new ActionResult();
-            setListener(emitter, result);
-            setPrintListener(emitter, result);
-            PlugPagTransactionResult plugPagTransactionResult = mPlugPag.doPayment(paymentData);
-            sendResponse(emitter, plugPagTransactionResult, result);
-        });
-    }
-
-    private void sendResponse(ObservableEmitter<ActionResult> emitter, PlugPagTransactionResult plugPagTransactionResult,
-                              ActionResult result) {
-        if (plugPagTransactionResult.getResult() != 0) {
-            emitter.onError(new PlugPagException(plugPagTransactionResult.getMessage(), plugPagTransactionResult.getErrorCode()));
-        } else {
-            result.setTransactionCode(plugPagTransactionResult.getTransactionCode());
-            result.setTransactionId(plugPagTransactionResult.getTransactionId());
-            emitter.onNext(result);
-        }
-        emitter.onComplete();
-    }
-
-    private void sendResponse(ObservableEmitter<ActionResult> emitter, PlugPagPrintResult printResult,
-                              ActionResult result) {
-
-        if (printResult.getResult() != 0) {
-            result.setResult(printResult.getResult());
-        }
-        emitter.onComplete();
-    }
-
-    private void setListener(ObservableEmitter<ActionResult> emitter, ActionResult result) {
-        mPlugPag.setEventListener(plugPagEventData -> {
-            result.setEventCode(plugPagEventData.getEventCode());
-            result.setMessage(plugPagEventData.getCustomMessage());
-            emitter.onNext(result);
-        });
-    }
-
-    private void setPrintListener(ObservableEmitter<ActionResult> emitter, ActionResult result) {
-        mPlugPag.setPrinterListener(new PlugPagPrinterListener() {
-            @Override
-            public void onError(PlugPagPrintResult printResult) {
-                result.setResult(printResult.getResult());
-                result.setMessage(String.format("Error %s %s", printResult.getErrorCode(), printResult.getMessage()));
-                result.setErrorCode(printResult.getErrorCode());
-                emitter.onNext(result);
-            }
-
-            @Override
-            public void onSuccess(PlugPagPrintResult printResult) {
-                result.setResult(printResult.getResult());
-                result.setMessage(String.format(Locale.getDefault(),"Print OK: Steps [%d]", printResult.getSteps()));
-                result.setErrorCode(printResult.getErrorCode());
-                emitter.onNext(result);
-            }
-        });
-    }
-
-    public PlugPagPaymentData getEventPaymentData(){
-        return mPlugPagPaymentData;
-    }
-
-    private int getAmount() {
-        return new Random().nextInt(10000) + 100;
-    }
-
-    private int getInstallments() {
-        return new Random().nextInt(5) + 1;
-    }
-
-    public Observable<ActionResult> printStablishmentReceipt() {
-        return Observable.create(emitter -> {
-            ActionResult actionResult = new ActionResult();
-            setPrintListener(emitter, actionResult);
-            PlugPagPrintResult result = mPlugPag.reprintStablishmentReceipt();
-            sendResponse(emitter, result, actionResult);
-        });
-    }
-
-    public Observable<ActionResult> printCustomerReceipt() {
-        return Observable.create(emitter -> {
-            ActionResult actionResult = new ActionResult();
-            setPrintListener(emitter, actionResult);
-            PlugPagPrintResult result = mPlugPag.reprintCustomerReceipt();
-            sendResponse(emitter, result, actionResult);
-        });
-    }
-
-    public Observable<Object> abort() {
-        return Observable.create(emitter -> {
-            mPlugPag.abort();
-            PlugPagAbortResult result = mPlugPag.abort();
-
-            if (result.getResult() == 0) {
-                emitter.onNext(new Object());
-            } else {
-                emitter.onError(new Exception("Erro ao abortar"));
-            }
-
-            emitter.onComplete();
-        });
-    }
-
-    public Observable<ActionResult> getLastTransaction() {
-        return Observable.create(emitter -> {
-            ActionResult actionResult = new ActionResult();
-
-            PlugPagTransactionResult result = mPlugPag.getLastApprovedTransaction();
-
-            sendResponse(emitter, result, actionResult);
-        });
-    }
-
-    public PlugPagCustomPrinterLayout getCustomPrinterDialog() {
-        PlugPagCustomPrinterLayout customDialog = new PlugPagCustomPrinterLayout();
-        customDialog.setTitle("Teste: Imprimir via do client?");
-        customDialog.setButtonBackgroundColor("#00ff33");
-        customDialog.setMaxTimeShowPopup(60);
-        return customDialog;
-    }
-}
