+++ app/src/main/java/br/com/uol/pagseguro/smartcoffee/utils/InstallmentConstants.java
@@ -2,10 +2,6 @@ package br.com.uol.pagseguro.smartcoffee.utils;
 
 public class InstallmentConstants {
 
-    public static final int INSTALLMENT_TYPE_A_VISTA = 1;
-    public static final int INSTALLMENT_TYPE_PARC_VENDEDOR = 2;
-    public static final int INSTALLMENT_TYPE_PARC_COMPRADOR = 3;
-
     public static final int INSTALLMENT_1X = 1;
     public static final String TOTAL_VALUE = "TOTAL_VALUE";
     public static final String INSTALLMENT_NUMBER = "INSTALLMENT_NUMBER";
