+++ app/src/main/res/drawable/ic_autentication.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#000000"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M7,24h2v-2L7,22v2zM11,24h2v-2h-2v2zM15,24h2v-2h-2v2zM16,0.01L8,0C6.9,0 6,0.9 6,2v16c0,1.1 0.9,2 2,2h8c1.1,0 2,-0.9 2,-2L18,2c0,-1.1 -0.9,-1.99 -2,-1.99zM16,16L8,16L8,4h8v12z"/>
+</vector>
