+++ app/src/main/res/layout/fragment_auth.xml
@@ -1,31 +1,34 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                                             xmlns:app="http://schemas.android.com/apk/res-auto"
-                                             android:layout_width="match_parent"
-                                             android:layout_height="match_parent"
-                                             android:background="@android:color/white"
-                                             android:clickable="true"
-                                             android:focusable="true">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@android:color/white"
+    android:clickable="true"
+    android:focusable="true">
 
     <Button
         android:id="@+id/btn_authentication_check"
-        style="@style/custom_button_style"
-        android:layout_height="wrap_content"
+        style="@style/ButtonStyle"
         android:text="@string/btn_authentication_check"
-        app:layout_constraintTop_toTopOf="parent"/>
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <Button
         android:id="@+id/btn_authentication_request"
-        style="@style/custom_button_style"
-        android:layout_height="wrap_content"
+        style="@style/ButtonStyle"
         android:text="@string/btn_authentication_request"
-        app:layout_constraintTop_toBottomOf="@+id/btn_authentication_check"/>
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/btn_authentication_check" />
 
     <Button
         android:id="@+id/btn_authentication_invalidate"
-        style="@style/custom_button_style"
-        android:layout_height="wrap_content"
+        style="@style/ButtonStyle"
         android:text="@string/btn_authentication_deactivate"
-        app:layout_constraintTop_toBottomOf="@+id/btn_authentication_request"/>
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/btn_authentication_request" />
 
 </android.support.constraint.ConstraintLayout>
