+++ app/src/main/res/layout/fragment_transactions.xml
@@ -14,66 +14,75 @@
 
         <Button
             android:id="@+id/btn_smartpos_credit"
-            style="@style/custom_button_style"
-            android:layout_height="wrap_content"
+            style="@style/ButtonStyle"
             android:text="@string/text_credit_in_cash"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
         <Button
             android:id="@+id/btn_smartpos_credit_with_seller_installments"
-            style="@style/custom_button_style"
-            android:layout_height="wrap_content"
+            style="@style/ButtonStyle"
             android:text="@string/text_credit_seller_installments"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/btn_smartpos_credit" />
 
         <Button
             android:id="@+id/btn_smartpos_credit_with_buyer_installments"
-            style="@style/custom_button_style"
-            android:layout_height="wrap_content"
+            style="@style/ButtonStyle"
             android:text="@string/text_credit_buyer_installments"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/btn_smartpos_credit_with_seller_installments" />
 
         <Button
             android:id="@+id/btn_smartpos_debit"
-            style="@style/custom_button_style"
-            android:layout_height="wrap_content"
+            style="@style/ButtonStyle"
             android:text="@string/text_debit"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/btn_smartpos_credit_with_buyer_installments" />
 
         <Button
             android:id="@+id/btn_smartpos_voucher"
-            style="@style/custom_button_style"
-            android:layout_height="wrap_content"
+            style="@style/ButtonStyle"
             android:text="@string/text_voucher"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/btn_smartpos_debit" />
 
         <Button
             android:id="@+id/btn_smartpos_void_payment"
-            style="@style/custom_button_style"
-            android:layout_height="wrap_content"
+            style="@style/ButtonStyle"
             android:text="@string/text_refund"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/btn_smartpos_voucher" />
 
         <Button
             android:id="@+id/btn_smartpos_void_print_stablishment"
-            style="@style/custom_button_style"
-            android:layout_height="wrap_content"
+            style="@style/ButtonStyle"
             android:text="@string/btn_smartpos_print_stablishment"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/btn_smartpos_void_payment" />
 
         <Button
             android:id="@+id/btn_smartpos_void_print_customer"
-            style="@style/custom_button_style"
-            android:layout_height="wrap_content"
+            style="@style/ButtonStyle"
             android:text="@string/btn_smartpos_print_customer"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/btn_smartpos_void_print_stablishment" />
 
 
         <Button
             android:id="@+id/btn_smartpos_get_last_transaction"
-            style="@style/custom_button_style"
-            android:layout_height="wrap_content"
+            style="@style/ButtonStyle"
             android:text="@string/btn_smartpos_last_transaction"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/btn_smartpos_void_print_customer" />
 
     </android.support.constraint.ConstraintLayout>
